refactor(migrate): fold allow-all ACL seed into 'migrate kafka' (drop separate subcommand)
