+++ cmd/migrate/kafka.go
@@ -28,6 +28,40 @@ func newKafkaCmd() *cobra.Command {
 	return cmd
 }
 
+// seedACLsEnabled gates seeding to the MSK path (SASL/SCRAM). It must NOT key
+// off UseSASL: OAUTHBEARER (GCP Managed Kafka) also sets UseSASL=true, and the
+// MSK-shaped "kafka-cluster"/User:* ACLs do not apply there.
+func seedACLsEnabled(m sarama.SASLMechanism) bool {
+	return m == sarama.SASLTypeSCRAMSHA256 || m == sarama.SASLTypeSCRAMSHA512
+}
+
+// seedAllowAllACLs seeds the allow-all ACL safety net using the SAME cluster
+// admin the topic reconcile already opened. Gated to SCRAM (MSK); a no-op on
+// any other mechanism. Idempotent — CreateACL on an existing binding is a
+// broker no-op. Never touches allow.everyone.if.no.acl.found.
+func seedAllowAllACLs(sa *reconcile.SaramaAdmin, mechanism sarama.SASLMechanism, dryRun bool) error {
+	if !seedACLsEnabled(mechanism) {
+		log.Printf("kafka-migrate: SASL mechanism %q is not SCRAM (MSK) — skipping ACL seed", mechanism)
+		return nil
+	}
+	rules := reconcile.AllowAllACLRules()
+	for _, r := range rules {
+		log.Printf("desired ACL: %v name=%q principal=%s op=%v", r.Resource.ResourceType, r.Resource.ResourceName, r.Acl.Principal, r.Acl.Operation)
+	}
+	if dryRun {
+		log.Printf("kafka-migrate: dry-run — %d allow-all ACL rules NOT applied", len(rules))
+		return nil
+	}
+	for _, r := range rules {
+		if err := sa.CreateACL(r.Resource, r.Acl); err != nil {
+			return fmt.Errorf("create ACL %v/%q: %w", r.Resource.ResourceType, r.Resource.ResourceName, err)
+		}
+		log.Printf("OK seeded ACL %v name=%q", r.Resource.ResourceType, r.Resource.ResourceName)
+	}
+	log.Printf("kafka-migrate: %d allow-all ACL rules ensured", len(rules))
+	return nil
+}
+
 func runKafkaMigration(dryRun bool) error {
 	cfg, err := config.NewConfig()
 	if err != nil {
@@ -82,6 +116,9 @@ func runKafkaMigration(dryRun bool) error {
 		for _, act := range plan {
 			logKafkaAction(act)
 		}
+		if err := seedAllowAllACLs(saramaAdmin, cfg.Kafka.SASLMechanism, true); err != nil {
+			return err
+		}
 		return nil
 	}
 
@@ -94,6 +131,9 @@ func runKafkaMigration(dryRun bool) error {
 	}
 	log.Printf("kafka-migrate done: created=%d grown=%d unchanged=%d skipped-shrink=%d rf-mismatch=%d retention-mismatch=%d",
 		res.Created, res.Grown, res.Unchanged, res.SkippedShrink, res.RFMismatch, res.RetentionMismatch)
+	if err := seedAllowAllACLs(saramaAdmin, cfg.Kafka.SASLMechanism, false); err != nil {
+		return err
+	}
 	return nil
 }
